Fix: removes props mutations (#352)
We are currently monitoring when data sources mutate props and found our
way to this repo because of
https://ops.grafana-ops.net/goto/mPuEDVPNg?orgId=1. These fixes should
remove any props mutation. I've tested this to the best of my abilities
but I'm no expert on this data source so I could use help testing this
so we keep the previous behavior.

https://lodash.info/doc/defaults

![image](https://github.com/user-attachments/assets/7c49262e-2944-4b57-9294-4ec2acc0c1b9)

---------

Co-authored-by: ivanahuckova <ivana.huckova@gmail.com>